requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.0
lxml==4.9.3
PyPDF2==3.0.1
python-docx==0.8.11
selenium==4.15.0
webdriver-manager==4.0.1
